Move SetServiceDefaults into the defaults pkg
