+++ operators/pkg/controller/common/defaults/service.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package common
+package defaults
 
 import (
 	v1 "k8s.io/api/core/v1"
