+++ operators/pkg/controller/common/defaults/service_test.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package common
+package defaults
 
 import (
 	"testing"
