+++ operators/pkg/controller/kibana/services.go
@@ -5,12 +5,13 @@
 package kibana
 
 import (
+	corev1 "k8s.io/api/core/v1"
+
 	kibanav1alpha1 "github.com/elastic/cloud-on-k8s/operators/pkg/apis/kibana/v1alpha1"
-	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/common"
+	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/common/defaults"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/kibana/label"
 	kbname "github.com/elastic/cloud-on-k8s/operators/pkg/controller/kibana/name"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/kibana/pod"
-	corev1 "k8s.io/api/core/v1"
 )
 
 func NewService(kb kibanav1alpha1.Kibana) *corev1.Service {
@@ -30,5 +31,5 @@ func NewService(kb kibanav1alpha1.Kibana) *corev1.Service {
 		},
 	}
 
-	return common.SetServiceDefaults(&svc, labels, labels, ports)
+	return defaults.SetServiceDefaults(&svc, labels, labels, ports)
 }
